requirements.txt
# Core dependencies
pydantic>=2.0.0
python-dotenv>=1.0.0
typing-extensions>=4.8.0

# Document processing
pypdf>=3.17.0
python-docx>=1.1.0
unstructured[pdf]>=0.11.0
pdfplumber>=0.10.3

# ML/AI frameworks
torch>=2.1.0
transformers>=4.35.0
scikit-learn>=1.3.0
langchain>=0.1.0
langchain-community>=0.0.10
llama-index>=0.9.0
sentence-transformers>=2.2.0

# Vector stores
psycopg2-binary>=2.9.0
pgvector>=0.2.3
chromadb>=0.4.18

# Testing
pytest>=7.4.3
pytest-cov>=4.1.0
pytest-asyncio>=0.21.1
pytest-mock>=3.12.0
hypothesis>=6.90.0
faker>=20.0.0

# Code quality
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.7.0
pylint>=3.0.0

# Monitoring and logging
prometheus-client>=0.19.0
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
structlog>=23.2.0

# Development tools
ipython>=8.18.0
jupyter>=1.0.0
pre-commit>=3.5.0

# API and web
fastapi>=0.104.0
uvicorn>=0.24.0
httpx>=0.25.0
aiofiles>=23.2.0

# Data manipulation
numpy>=1.24.0
pandas>=2.1.0
pyarrow>=14.0.0

# Configuration
pyyaml>=6.0.1
toml>=0.10.2